Clean and Simplify code further
evaluateEqualsInput function was redundant by cleaning up the
operate function and realizing the if condition was not neeeded.

Other simplifications of the code largely around replacing redundant
logic in if/else statements that could be handled by 1 ternary
operator at the end. Again helped identify these areas with AI and
then verified the functionality.

Used as a practice to get better at making efficient use of the ternary
operator.